Task3_Leaf_Disease_Detection/requirements.txt
tensorflow>=2.10.0
keras>=2.10.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
Pillow>=8.3.0
opencv-python>=4.5.0
plotly>=5.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
tqdm>=4.62.0
